aws-serverless/cdk/requirements.txt
# CDK Requirements for Financial AI Quality Enhancement API
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.100.0a0

# Core Python dependencies
boto3>=1.34.0
botocore>=1.34.0

# Financial data APIs
yfinance>=0.2.28
pandas>=2.1.4
numpy>=1.26.2

# HTTP and JSON processing
requests>=2.31.0

# Date and time handling
python-dateutil>=2.8.2

# Logging
structlog>=23.2.0
